requirements.txt
aiofiles==22.1.0
aiosmtplib==1.1.7
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
autopep8==2.0.0
bcrypt==3.2.2
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cryptography==37.0.4
cssselect==1.2.0
cssutils==2.6.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
emails==0.6
fastapi==0.78.0
greenlet==2.0.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
orjson==3.7.12
passlib==1.7.4
Pillow==9.3.0
premailer==3.10.0
psycopg==3.0.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.6
urllib3==1.26.12
uvicorn==0.18.2